+++ PGUPV/src/window.cpp
@@ -472,7 +472,7 @@ void Window::setTitle(const std::string &title) {
 	window->setTitle(title);
 }
 
-void Window::setMousePosition(uint x, uint y) {
+void Window::setMousePosition(float x, float y) {
 	assert(window);
 	window->setMousePosition(x, y);
 }
